+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Release Notes:
 
 ### vNext
-
+- [DAV-3719] Fix snackbar tour error in console
 - [DAV-3792] Align asset transfer UI with the design
 - [DAV-3734] Add in-wallet transaction modals
 - [DAV-3669] Add asset transfer (ethereum)